fix sky model
