+++ src/models/index.jsx
@@ -1,0 +1,4 @@
+export { BooIsland } from "./BooIsland/BooIsland";
+export { Fox } from "./Fox/Fox";
+export { Plane } from "./Plane/Plane";
+export { Sky } from "./Sky/Sky";
